Allow variance in task execution strategy
Why
---

The host application _may_ want more control over how Shipit schedules
and executes work given the topology / capabilities of its environment.

An example, _we_ deploy our host application into a Kubernetes cluster.
We'd like the ability to schedule Shipit::Deploy tasks to run in their
own pod, rather than relying on a processing the Shipit::Deploy task
inside the context of the Sidekiq worker container - which _may_ be
processing other jobs, like Sidekiq health checks, etc.

How
---

This change set extracts the current task execution logic into a
"Default Execution Strategy" - this change should NOT affect a vanilla
instance of shipit-engine. We also expose an interface whereby the host
application _may_ register a custom "execution strategy" for a given
task type - overriding the default. The `PerformTaskJob` looks-up the
strategy to use for a given task, then executes the task using the
appropriate execution strategy.

References
----------

- https://github.com/Shopify/shipit-engine/pull/1107
